Move `git_attr_value()` bindings to a submodule
